+++ requirements/base-requirements.txt
@@ -1,3 +1,4 @@
+# Others
 flask-marshmallow==0.14.0
 Flask==2.2.2
 marshmallow==3.19.0
