+++ requirements/requirements.txt
@@ -1,0 +1,1 @@
+-r base-requirements.txt
